# sdks/sdkwork-im-backend-sdk/openapi/sdkwork-im-backend-api.openapi.yaml
openapi: 3.1.2
info:
  title: Sdkwork IM Backend Management API
  version: 0.1.0
  description: Backend management OpenAPI contract for operator, governance, control, and admin APIs under one SDK family.
servers:
  - url: http://127.0.0.1:18079
security:
  - AuthToken: []
    AccessToken: []
...
paths:
...
  /backend/v3/api/ops/cluster:
    get:
      tags:
        - ops
      operationId: cluster.retrieve
      summary: Retrieve cluster state
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SdkWorkApiResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        type: object
                        required:
                          - item
                        properties:
                          item:
                            type: object
                            additionalProperties: true
        '401':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - AuthToken: []
          AccessToken: []
      x-sdkwork-owner: sdkwork-im
      x-sdkwork-api-authority: sdkwork-im.backend
      x-sdkwork-request-context: WebRequestContext
      x-sdkwork-api-surface: backend-api
  /backend/v3/api/ops/lag:
    get:
      tags:
        - ops
      operationId: lag.retrieve
      summary: Retrieve operational lag
      parameters:
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/CursorQuery'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LagListResponse'
        '400':
          description: Invalid pagination query or cursor
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - AuthToken: []
          AccessToken: []
      x-sdkwork-owner: sdkwork-im
      x-sdkwork-api-authority: sdkwork-im.backend
      x-sdkwork-request-context: WebRequestContext
      x-sdkwork-api-surface: backend-api
...
  /backend/v3/api/ops/runtime_dir:
    get:
      tags:
        - ops
      operationId: runtimeDir.retrieve
      summary: Inspect runtime directory
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SdkWorkApiResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        type: object
                        required:
                          - item
                        properties:
                          item:
                            type: object
                            additionalProperties: true
        '401':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - AuthToken: []
          AccessToken: []
      x-sdkwork-owner: sdkwork-im
      x-sdkwork-api-authority: sdkwork-im.backend
      x-sdkwork-request-context: WebRequestContext
      x-sdkwork-api-surface: backend-api
  /backend/v3/api/ops/provider_bindings:
    get:
      tags:
        - ops
      operationId: ops.providerBindings.list
      summary: List provider bindings
      parameters:
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/CursorQuery'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderBindingSnapshotListResponse'
        '400':
          description: Invalid pagination query or cursor
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - AuthToken: []
          AccessToken: []
      x-sdkwork-owner: sdkwork-im
      x-sdkwork-api-authority: sdkwork-im.backend
      x-sdkwork-request-context: WebRequestContext
      x-sdkwork-api-surface: backend-api
  /backend/v3/api/ops/provider_bindings/drift:
    get:
      tags:
        - ops
      operationId: ops.providerBindings.drift.retrieve
      summary: Retrieve provider binding drift
      parameters:
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/CursorQuery'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderBindingDriftListResponse'
        '400':
          description: Invalid pagination query or cursor
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - AuthToken: []
          AccessToken: []
      x-sdkwork-owner: sdkwork-im
      x-sdkwork-api-authority: sdkwork-im.backend
      x-sdkwork-request-context: WebRequestContext
      x-sdkwork-api-surface: backend-api
  /backend/v3/api/ops/diagnostics:
    get:
      tags:
        - ops
      operationId: diagnostics.retrieve
      summary: Retrieve diagnostics
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SdkWorkApiResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        type: object
                        required:
                          - item
                        properties:
                          item:
                            type: object
                            additionalProperties: true
        '401':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - AuthToken: []
          AccessToken: []
      x-sdkwork-owner: sdkwork-im
      x-sdkwork-api-authority: sdkwork-im.backend
      x-sdkwork-request-context: WebRequestContext
      x-sdkwork-api-surface: backend-api
  /backend/v3/api/ops/replay_status:
    get:
      tags:
        - ops
      operationId: replayStatus.retrieve
      summary: Retrieve commit-journal replay status
      description: >-
        Reports the commit-journal replay surface: whether the shared
        PostgreSQL journal pool is configured and the journal head cursor
        (committed offset and row count). Returns a real status view instead
        of 404 when the journal is not configured in this service.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SdkWorkApiResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        type: object
                        required:
                          - item
                        properties:
                          item:
                            $ref: '#/components/schemas/JournalReplayStatusView'
        '401':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - AuthToken: []
          AccessToken: []
      x-sdkwork-owner: sdkwork-im
      x-sdkwork-api-authority: sdkwork-im.backend
      x-sdkwork-request-context: WebRequestContext
      x-sdkwork-api-surface: backend-api
  /backend/v3/api/ops/retention/purge:
    post:
      tags:
        - ops
      operationId: retention.purge
      summary: Purge expired retention batches
      description: >-
        Runs a bounded retention purge batch over expired commit-journal and
        derived rows. Optional `batch_size` query parameter bounds the batch
        (default 500, maximum 5000).
      parameters:
        - name: batch_size
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
            maximum: 5000
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SdkWorkApiResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        type: object
                        required:
                          - item
                        properties:
                          item:
                            $ref: '#/components/schemas/RetentionPurgeResponse'
        '401':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          description: Problem detail response
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - AuthToken: []
          AccessToken: []
      x-sdkwork-owner: sdkwork-im
      x-sdkwork-api-authority: sdkwork-im.backend
      x-sdkwork-request-context: WebRequestContext
      x-sdkwork-api-surface: backend-api
...
components:
  securitySchemes:
    AuthToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AccessToken:
      type: apiKey
      in: header
      name: Access-Token
...
  parameters:
    PageSizeQuery:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 20
    CursorQuery:
      name: cursor
      in: query
      required: false
      schema:
        type: string
...
  schemas:
    ProblemDetail:
      type: object
      additionalProperties: true
      required:
        - type
        - title
        - status
        - code
        - traceId
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        code:
          $ref: '#/components/schemas/SdkWorkPlatformErrorCode'
        traceId:
          type: string
          format: uuid
          description: Server-owned request correlation id.
        i18nKey:
          type: string
          description: Optional stable localization key such as errors.result.40001.
        locale:
          type: string
          description: Optional effective BCP 47 locale used by framework message mapping.
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
...
    JournalReplayStatusView:
      type: object
      additionalProperties: false
      required:
        - status
        - mode
        - databaseConfigured
        - journalReady
        - generatedAt
      properties:
        status:
          type: string
          enum:
            - enabled
            - not_configured
            - unavailable
        mode:
          type: string
          enum:
            - postgres-journal
            - unconfigured
        databaseConfigured:
          type: boolean
        journalReady:
          type: boolean
        totalCommits:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        headCommitOffset:
          type: integer
          format: int64
          nullable: true
        latestOccurredAt:
          type: string
          format: date-time
          nullable: true
        detail:
          type: string
          nullable: true
        generatedAt:
          type: string
          format: date-time
    RetentionPurgeResponse:
      type: object
      additionalProperties: false
      required:
        - generatedAt
        - batchSize
      properties:
        generatedAt:
          type: string
          format: date-time
        batchSize:
          type: integer
          format: int64
        commitJournalDeleted:
          type: integer
          format: int64
          minimum: 0
        conversationMessagesDeleted:
          type: integer
          format: int64
          minimum: 0
        messageMediaRefsDeleted:
          type: integer
          format: int64
          minimum: 0
        outboxEventsDeleted:
          type: integer
          format: int64
          minimum: 0
        inboxEventsDeleted:
          type: integer
          format: int64
          minimum: 0
        realtimeDeviceEventsDeleted:
          type: integer
          format: int64
          minimum: 0
        rtcSessionsDeleted:
          type: integer
          format: int64
          minimum: 0
        rtcSignalsDeleted:
          type: integer
          format: int64
          minimum: 0
...
    LagItem:
      type: object
      additionalProperties: false
      required:
        - component
        - scopeId
        - currentOffset
        - committedOffset
        - lag
      properties:
        component:
          type: string
        scopeId:
          type: string
        currentOffset:
          type: string
          format: uint64
          pattern: ^[0-9]+$
          x-sdkwork-int64-string: true
          x-sdkwork-rust-type: u64
        committedOffset:
          type: string
          format: uint64
          pattern: ^[0-9]+$
          x-sdkwork-int64-string: true
          x-sdkwork-rust-type: u64
        lag:
          type: string
          format: uint64
          pattern: ^[0-9]+$
          x-sdkwork-int64-string: true
          x-sdkwork-rust-type: u64
    ProviderBindingItem:
      type: object
      additionalProperties: false
      required:
        - domain
        - defaultPluginId
        - selectedPluginId
        - selectionSource
        - tenantOverrideAllowed
      properties:
        domain:
          type: string
        defaultPluginId:
          type:
            - string
            - 'null'
        selectedPluginId:
          type:
            - string
            - 'null'
        selectionSource:
          type: string
        tenantOverrideAllowed:
          type: boolean
    ProviderBindingSnapshot:
      type: object
      additionalProperties: false
      required:
        - interfaceVersion
        - tenantId
        - effectiveBindings
        - precedence
      properties:
        interfaceVersion:
          type: string
        tenantId:
          type:
            - string
            - 'null'
        effectiveBindings:
          type: array
          items:
            $ref: '#/components/schemas/ProviderBindingItem'
        precedence:
          type: array
          items:
            type: string
    ProviderBindingDriftItem:
      type: object
      additionalProperties: false
      required:
        - tenantId
        - domain
        - baselineSelectedPluginId
        - selectedPluginId
        - baselineSelectionSource
        - selectionSource
        - driftKind
      properties:
        tenantId:
          type: string
        domain:
          type: string
        baselineSelectedPluginId:
          type:
            - string
            - 'null'
        selectedPluginId:
          type:
            - string
            - 'null'
        baselineSelectionSource:
          type: string
        selectionSource:
          type: string
        driftKind:
          type: string
    LagPageData:
      type: object
      additionalProperties: false
      required:
        - items
        - pageInfo
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LagItem'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
    ProviderBindingSnapshotPageData:
      type: object
      additionalProperties: false
      required:
        - items
        - pageInfo
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProviderBindingSnapshot'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
    ProviderBindingDriftPageData:
      type: object
      additionalProperties: false
      required:
        - items
        - pageInfo
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProviderBindingDriftItem'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
    LagListResponse:
      allOf:
        - $ref: '#/components/schemas/SdkWorkApiResponse'
        - type: object
          required:
            - data
          properties:
            data:
              $ref: '#/components/schemas/LagPageData'
    ProviderBindingSnapshotListResponse:
      allOf:
        - $ref: '#/components/schemas/SdkWorkApiResponse'
        - type: object
          required:
            - data
          properties:
            data:
              $ref: '#/components/schemas/ProviderBindingSnapshotPageData'
    ProviderBindingDriftListResponse:
      allOf:
        - $ref: '#/components/schemas/SdkWorkApiResponse'
        - type: object
          required:
            - data
          properties:
            data:
              $ref: '#/components/schemas/ProviderBindingDriftPageData'
...
    SdkWorkApiResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - data
        - traceId
      properties:
        code:
          type: integer
          format: int32
          enum:
            - 0
          default: 0
          minimum: 0
          maximum: 0
        data:
          description: Operation-specific payload typed per response schema.
        traceId:
          type: string
          format: uuid
          description: Server-owned request correlation id.
...
    PageInfo:
      type: object
      additionalProperties: false
      required:
        - mode
      properties:
        mode:
          type: string
          enum:
            - offset
            - cursor
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 200
        totalItems:
          type: string
          pattern: ^[0-9]+$
        totalPages:
          type: integer
          minimum: 0
        nextCursor:
          type:
            - string
            - 'null'
        hasMore:
          type: boolean
    SdkWorkPlatformErrorCode:
      type: integer
      format: int32
      minimum: 40001
      maximum: 79999
      description: Platform or domain error code per API_SPEC.md §15.3.
    FieldError:
      type: object
      additionalProperties: false
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
        code:
          type: integer
          format: int32
          minimum: 40011
          maximum: 40099
        i18nKey:
          type: string
        params:
          type: object
          additionalProperties:
            oneOf:
              - type: string
              - type: number
              - type: integer
              - type: boolean
...